feat: put all filters together in the sidebar
